realizamos los ultimos cambios para la entrega en las modificaciones de las sesiones

+++ admin/abml/baja.php
@@ -46,7 +46,7 @@
         }
     }
     if (isset($_GET["idS"])) {
-        $idSesion = htmlspecialchars($_GET["idS"]);
+        $idSesion = (int) htmlspecialchars($_GET["idS"]);
         mysqli_query($conexion,"DELETE FROM `sesiones` WHERE `id_sesiones`='$idSesion'");
         mysqli_query($conexion,"DELETE FROM `sesiones_tratamientos` WHERE `fk_sesiones`='$idSesion'");
         mysqli_query($conexion,"DELETE FROM `pago_sesiones` WHERE `fk_sesiones`='$idSesion'");
